Add 'start' and 'stop' timeouts to control AWS OpsWorks instances.
The 10 minute timeout was found to be too short for the OpsWorks
instances we were creating and destroying. This meant that the
terraform operations would regularly time out, and necessitated
changes to the code for the OpsWorks timeouts.

This change introduces two new Timeout types - 'Start' and 'Stop' -
for the time needed to start or stop the instance. Other terraform
rules have their timeouts in the 'Create', 'Destroy' or 'Update'
rules. The start and stop operations may need very different timeouts
and if these operations were mapped directly to 'Create', 'Destroy'
and 'Update', then they would be forced to be the start instance
time, the stop instance time, and the greater of 'Start' and 'Stop',
which would not be especially obvious.

Thus, I have decided to introduce new types of timeouts for this
case.