Update all the existing tests to work with the new Timeouts struct.
The new struct has 2 additional members - start and stop - which need
to be tested in addition to the existing regression tests. Those tests
that check the behaviour of the timeouts needs to be able to check
these parameters with the others. The functions that manipulate
and check the timeouts have the new start and stop fields added.

This change makes the existing tests work with those new parameters.
As a side effect, we do test the 'default' parameter works against
stop and start, but a further change will add extra tests for the
stop and start fields.